Add function to copy iPXE binaries
.gitlab/ci/build_releng.sh:
Add `copy_ipxe_binaries()` to copy the relevant iPXE binaries to an ipxe directory in the output directory.
Change `run_mkarchiso()` to also call `copy_ipxe_binaries()` and create checksums for the iPXE binaries.
Change `create_metrics()` to create info metrics for the version of archiso used for the build, the kernel version used
in the created image and the ipxe version used when exporting the ipxe binaries.