Replace local gitlab-ci include with ci-scripts
.gitlab-ci.yml:
Replace the local include of a gitlab-ci description (and the accompanying script) with an include of
archlinux/ci-scripts.
Copy potentially existing keys in script target instead of before_script target. This way we get to remove before_script
and only need to specify which script from ci-scripts we are calling in the scripts target, instead of redoing all of
before_script.
Remove after_script target, as it is taken care of in ci-scripts.